Moved List component into separate directory
