+++ src/List/List.js
@@ -1,0 +1,3 @@
+import basicClassCreator from '../utils/basicClassCreator';
+
+export default basicClassCreator('List', 'mdl-list', 'ul');
