+++ src/List/__tests__/List-test.js
@@ -1,0 +1,13 @@
+/* eslint-env mocha */
+import expect from 'expect';
+import React from 'react';
+import { render } from '../../__tests__/render';
+import List from '../List';
+
+describe('List', () => {
+    it('should render an ul list', () => {
+        const output = render(<List />);
+
+        expect(output.type).toBe('ul');
+    });
+});
